backend/requirements.txt
# Finto GST Reconciliation Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Supabase
supabase>=2.3.0

# File Processing
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0

# AI - Groq
groq>=0.4.0

# Background Jobs
celery>=5.3.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
